Productionise code, split huge files and remove outputs
