+++ Makefile
@@ -2,14 +2,32 @@ all: cube
 
 OBJS = build/parser.o  \
 	build/codegen.o \
+	build/linker_common.o \
 	build/main.o    \
 	build/tokens.o  \
 	build/core.o  \
+	build/nassignment.o \
+	build/nbinaryoperator.o \
+	build/nblock.o \
+	build/nconstdeclaration.o \
+	build/ndouble.o \
+	build/nexpressionstatement.o \
+	build/nfunction.o \
+	build/nidentifier.o \
+	build/ninteger.o \
+	build/ninternaldeclaration.o \
+	build/nmethodcall.o \
+	build/nprogram.o \
+	build/nreturn.o \
+	build/nstring.o \
+	build/nsuperdeclaration.o \
+	build/nthing.o \
+	build/nvariabledeclaration.o
 
 LLVMCONFIG = llvm-config
 CPPFLAGS = `$(LLVMCONFIG) --cppflags` -std=c++20 -lstdc++fs
 LDFLAGS = `$(LLVMCONFIG) --ldflags` -lpthread -ldl -lz -lncurses
-LIBS = `$(LLVMCONFIG) --system-libs --libs all`
+LIBS = `$(LLVMCONFIG) --system-libs --libs all` -L./build/ -L.
 
 .PHONY: clean
 clean:
@@ -31,5 +49,5 @@ build/tokens.cpp: build/tokens.l build/parser.hpp
 	g++ -c $(CPPFLAGS) -o $@ $<
 
 
-cube: build/ $(OBJS)
+cube: build/ $(COMMON) $(OBJS)
 	g++ -o $@ $(OBJS) $(LIBS) $(LDFLAGS)
