+++ src/linker_common.h
@@ -1,0 +1,3 @@
+#include "node.h"
+
+static llvm::Type* typeOf(NIdentifier& type);
