+++ src/nconstdeclaration.cpp
@@ -1,0 +1,10 @@
+#include "codegen.h"
+#include "node.h"
+
+// NConstDeclaration::codeGen does nothing for now
+//
+// TODO: implementation
+Value* NConstDeclaration::codeGen(CodeGenContext& context)
+{
+  return NULL;
+}
