+++ src/ninternaldeclaration.cpp
@@ -1,0 +1,10 @@
+#include "codegen.h"
+#include "node.h"
+
+// NInternalDeclaration::codeGen does nothing for now
+//
+// TODO: implementation
+Value* NInternalDeclaration::codeGen(CodeGenContext& context)
+{
+  return NULL;
+}
