+++ src/nsuperdeclaration.cpp
@@ -1,0 +1,10 @@
+#include "codegen.h"
+#include "node.h"
+
+// NSuperDeclaration::codeGen does nothing for now
+//
+// TODO: implementation
+Value* NSuperDeclaration::codeGen(CodeGenContext& context)
+{
+  return NULL;
+}
