+++ src/nthing.cpp
@@ -1,0 +1,10 @@
+#include "codegen.h"
+#include "node.h"
+
+// NThingDeclaration does nothing for now.
+//
+// TODO: implement
+Value* NThingDeclaration::codeGen(CodeGenContext& context)
+{
+  return NULL;
+}
